altera models para incluir campo que aceita decimais

+++ sport_ml_djavue/sport_ml_djavue/sport_predicts/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Data(models.Model):
     name = models.CharField(max_length=100, blank=False)
     age = models.PositiveIntegerField(blank=False)
-    height = models.FloatField(blank=False)
+    height = models.DecimalField(max_digits=3, decimal_places=2, blank=False)
     sex = models.IntegerField(blank=False)
     predictions = models.CharField(max_length=30, blank=False)
     date = models.DateTimeField(auto_now_add=True)
